+++ CHANGELOG.md
@@ -10,7 +10,7 @@
   - configs now take effect without the need for a restart!
   - in simple mode, the template inventory is now virtual: it doesn't actually contain items. It remembers and displays the items you click on it with, but you don't lose those items.
   - because of the change listed above **you will lose any items** placed in the template area *in previous versions*. If you want to keep your template items, you'll need to manually remove them **before updating**. 
-  - in simple mode, players can now place items into the input area if they match the template
+  - in simple mode, players can now place items into the input area if they match the template+
 
 - 1.3.24 (9 Jul. 2021):